Refact: Fixed bugs and update services

+++ src/main/java/com/smarterfit/modules/classgroup/event/listener/ClassGroupEventListener.java
@@ -18,7 +18,9 @@ public class ClassGroupEventListener {
 
     @EventListener
     public void handleClassGroupDeactivated(ClassGroupDeactivatedEvent event) {
-
+        if (event == null || event.getClassGroup() == null) {
+            return;
+        }
         classGroupUserService.removeSubscriptionByClassGroup(event.getClassGroup());
     }
 }
